Extract waveform signal widget module
